+++ .php-cs-fixer.dist.php
@@ -83,7 +83,6 @@ $files = [
 
 $exclude = [
     'fonts',
-    'Legacy',
     'resources',
 ];