+++ tests/PdfCompareBasicTest.php
@@ -25,7 +25,6 @@ use fpdf\Enums\PdfZoom;
 use fpdf\PdfBorder;
 use fpdf\PdfDocument;
 use fpdf\PdfException;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareBasicTest extends AbstractCompareTestCase
 {
@@ -109,7 +108,7 @@ final class PdfCompareBasicTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $doc->SetCompression(false);
         $doc->SetFont('Arial', 'B', 16);