+++ tests/PdfCompareBorderTest.php
@@ -17,7 +17,6 @@ use fpdf\Enums\PdfMove;
 use fpdf\PdfBorder;
 use fpdf\PdfDocument;
 use fpdf\PdfException;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareBorderTest extends AbstractCompareTestCase
 {
@@ -42,7 +41,7 @@ final class PdfCompareBorderTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $width = 0.0;
         $height = 5.0;