+++ tests/PdfCompareEscape.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareEscape extends AbstractCompareTestCase
 {
@@ -28,7 +27,7 @@ final class PdfCompareEscape extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $text = $this->getText();
         $doc->Text(5.0, 15.0, $text);