+++ tests/PdfCompareImageTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareImageTest extends AbstractCompareTestCase
 {
@@ -29,7 +28,7 @@ final class PdfCompareImageTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $doc->Image(__DIR__ . '/images/image.png');
         $doc->Ln(5.0);