+++ tests/PdfCompareLineTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareLineTest extends AbstractCompareTestCase
 {
@@ -27,7 +26,7 @@ final class PdfCompareLineTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $doc->Line(10, 10, 100, 100);
         $doc->SetLineWidth(1.0);