+++ tests/PdfCompareMetadataTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareMetadataTest extends AbstractCompareTestCase
 {
@@ -29,7 +28,7 @@ final class PdfCompareMetadataTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $doc->SetTitle('Title æ', true);
         $doc->SetSubject('Subject');