+++ tests/PdfCompareOrientationTest.php
@@ -18,7 +18,6 @@ use fpdf\Enums\PdfPageSize;
 use fpdf\Enums\PdfRotation;
 use fpdf\PdfDocument;
 use fpdf\PdfSize;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareOrientationTest extends AbstractCompareTestCase
 {
@@ -57,7 +56,7 @@ final class PdfCompareOrientationTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $doc->Cell(0.0, 0.0, 'Portrait');