+++ tests/PdfCompareTextTest.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace fpdf\Tests;
 
 use fpdf\PdfDocument;
-use fpdf\Tests\Legacy\FPDF;
 
 final class PdfCompareTextTest extends AbstractCompareTestCase
 {
@@ -25,7 +24,7 @@ final class PdfCompareTextTest extends AbstractCompareTestCase
     }
 
     #[\Override]
-    protected function updateOldDocument(FPDF $doc): void
+    protected function updateOldDocument(\FPDF $doc): void
     {
         $doc->Text(5.0, 15.0, "This is a text test.\nWith multi-lines.");
     }